Better implementation of keyvalues.
-Fixed wrong fade in ammo bar.
-Added a better and cleaner keyvalues implementation( No longer crashes when the game is in closing state ).
-Better auto pistol implementation.
