+++ README.md
@@ -9,7 +9,6 @@ If you want a clean base without any features, please see the [deadcell SDK](htt
 - Crashes and lots of them.
 - Game will crash when using -d3d9ex as a launch option.
 - Lots of code is unoptimized.
-- Visuals will not rescale if you change resolution while injected.
 - If you come across any more issues, please notify us via the issue tracker.
 
 We are accepting any pull requests that fix any of the aforementioned issues.
